feat: add delete request handling
