+++ src/server.ts
@@ -1,7 +1,52 @@
 import http, { IncomingMessage, ServerResponse } from 'http';
-import { validate, v4 as uuidv4 } from 'uuid';
-import { UserData, users } from './storage/users';
+import { validate, v4 as uuidv4, UUIDTypes } from 'uuid';
 import { validateUserData } from './utils/validateUserData';
+
+export type UserData = {
+  id: UUIDTypes
+  username: string
+  age: number
+  hobbies: string[] | []
+}
+
+let users: UserData[] = [
+  {
+    username: 'johndoe',
+    id: uuidv4(),
+    age: 27,
+    hobbies: ['playing guitar', 'sining', 'gym']
+  },
+  {
+    username: 'janedoe',
+    id: uuidv4(),
+    age: 42,
+    hobbies: ['oil painting', 'stretching']
+  },
+  {
+    username: 'roten_tomato',
+    id: uuidv4(),
+    age: 42,
+    hobbies: ['cinematography', 'sining', 'swimming']
+  },
+  {
+    username: 'fluffyemokitty',
+    id: uuidv4(),
+    age: 18,
+    hobbies: ['sining', 'dancing']
+  },
+  {
+    username: 'unique_diamond',
+    id: uuidv4(),
+    age: 20,
+    hobbies: ['dancing']
+  },
+  {
+    username: 'streetwalker',
+    id: uuidv4(),
+    age: 60,
+    hobbies: []
+  }
+];
 
 const USERS_ROUTE = '/api/users';
 
@@ -135,6 +180,34 @@ const server = http.createServer((request: IncomingMessage, response: ServerResp
             response.end();
           }
         })
+      }
+    } else if (method === 'DELETE') {
+      if (url?.startsWith(`${USERS_ROUTE}/`)) {
+        const urlPathname = url.split('/');
+        const userID = urlPathname[urlPathname.length - 1];
+        const isIdValid = validate(userID);
+        if (!isIdValid) {
+          response.statusCode = 400;
+          response.statusMessage = `User id: ${userID} is invalid`;
+    
+          response.end();
+        }
+
+        if (userID && isIdValid) {
+          const user = users.find(userData => userData.id === userID);
+          if (user) {
+            users = users.filter(user => user.id !== userID)
+            response.setHeader('Content-Type', 'application/json');
+            response.statusCode = 204;
+      
+            response.end(JSON.stringify(user));
+          } else {
+            response.statusCode = 404;
+            response.statusMessage = `User with id: ${userID} doesn't exist`;
+      
+            response.end();
+          }
+        }
       }
     } else {
       response.statusCode = 404;
